Solve issue 48: Python runtime deadlocks when a large amount of data is returned
The current Python runtime for Kubeless will deadlock when more data than the
OS pipe can handle is pushed to the queue.  The server blocks while the process
is alive and only then attempts to get the response from the queue.  However,
the process stays alive because it cannot push more data to the queue unless it
is read from.

This update instead blocks on reading from the queue and receives data as it is
pushed.